requirements.txt

streamlit
torch
torchvision
yolov5
opencv-python-headless
pytesseract
matplotlib
numpy
boto3
pytesseract
Pillow
pydantic
python-dotenv
google-generativeai
langgraph
langchain
langchain-core
gspread
google-auth
openpyxl
pandas
